pedidos recibidos

+++ app/src/main/java/com/grupo6/trego/data/model/DTOProductoPedido.kt
@@ -5,7 +5,7 @@ data class DTOProductoPedido(
     val ingredientes: List<DTOIngrediente>?,
     val observaciones: String?,
     val cantidad: Int?,
-    val subtotal: Double?,
+    val subtotal: Float?,
     val producto: DTOProductoSimplificado?
 ) {
 }
